# openapi.yaml
openapi: 3.0.3
info:
  title: Algorithmic Trading Platform API
  version: 0.1.0
  description: >
    API contract for a modular algorithmic trading platform supporting market data ingestion,
    feature computation, backtesting, paper/live runs, risk enforcement, and execution.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /control/trading/halt:
    post:
      tags: [Control]
      summary: Halt trading (kill switch)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlRequest"
      responses:
        "200":
          description: Trading halted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /control/trading/resume:
    post:
      tags: [Control]
      summary: Resume trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlRequest"
      responses:
        "200":
          description: Trading resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /marketdata/bars:
    get:
      tags: [MarketData]
      summary: Fetch historical bars from internal storage
      parameters:
        - $ref: "#/components/parameters/symbol"
        - $ref: "#/components/parameters/timeframe"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Bars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /marketdata/backfill:
    post:
      tags: [MarketData, Jobs]
      summary: Trigger historical backfill job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillRequest"
      responses:
        "202":
          description: Backfill job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnqueuedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /features/compute:
    post:
      tags: [Features, Jobs]
      summary: Compute and store features (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureComputeRequest"
      responses:
        "202":
          description: Feature compute job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnqueuedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /features:
    get:
      tags: [Features]
      summary: Fetch computed features
      parameters:
        - $ref: "#/components/parameters/symbol"
        - $ref: "#/components/parameters/timeframe"
        - name: feature_set
          in: query
          required: false
          schema:
            type: string
            default: basic_v1
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: Feature rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /strategies:
    get:
      tags: [Strategies]
      summary: List available strategies
      responses:
        "200":
          description: Strategies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategiesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /strategies/{strategy_id}:
    get:
      tags: [Strategies]
      summary: Get strategy detail and parameter schema
      parameters:
        - $ref: "#/components/parameters/strategy_id"
      responses:
        "200":
          description: Strategy detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    symbol:
      name: symbol
      in: query
      required: true
      schema:
        type: string
    timeframe:
      name: timeframe
      in: query
      required: true
      schema:
        type: string
        enum: [1m, 5m, 1h, 1d]
    start:
      name: start
      in: query
      required: true
      schema:
        type: string
        format: date-time
    end:
      name: end
      in: query
      required: true
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 500
...
    strategy_id:
      name: strategy_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_ARGUMENT
        message:
          type: string
          example: symbol is required
        details:
          type: object
          additionalProperties: true
...
    ControlRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ControlResponse:
      type: object
      required: [halted, time]
      properties:
        halted:
          type: boolean
        time:
          type: string
          format: date-time

    # ---- Symbols ----
...
    Bar:
      type: object
      required: [ts, open, high, low, close, volume]
      properties:
        ts:
          type: string
          format: date-time
        open:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        close:
          type: number
          format: double
        volume:
          type: integer
          format: int64
    BarsResponse:
      type: object
      required: [symbol, timeframe, bars]
      properties:
        symbol:
          type: string
        timeframe:
          type: string
          enum: [1m, 5m, 1h, 1d]
        bars:
          type: array
          items:
            $ref: "#/components/schemas/Bar"
    BackfillRequest:
      type: object
      required: [symbols, timeframe, start, end, source]
      properties:
        symbols:
          type: array
          items: { type: string }
          minItems: 1
        timeframe:
          type: string
          enum: [1m, 5m, 1h, 1d]
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        source:
          type: string
          example: polygon

    # ---- Jobs ----
    JobEnqueuedResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued]
...
    FeatureComputeRequest:
      type: object
      required: [symbols, timeframe, start, end, feature_set]
      properties:
        symbols:
          type: array
          items: { type: string }
          minItems: 1
        timeframe:
          type: string
          enum: [1m, 5m, 1h, 1d]
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        feature_set:
          type: string
          example: basic_v1
    FeatureRow:
      type: object
      required: [ts]
      properties:
        ts:
          type: string
          format: date-time
      additionalProperties: true
    FeaturesResponse:
      type: object
      required: [symbol, timeframe, feature_set, rows]
      properties:
        symbol:
          type: string
        timeframe:
          type: string
          enum: [1m, 5m, 1h, 1d]
        feature_set:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/FeatureRow"

    # ---- Strategies ----
    StrategyMeta:
      type: object
      required: [strategy_id, name, type]
      properties:
        strategy_id:
          type: string
          example: sma_cross_v1
        name:
          type: string
          example: SMA Crossover v1
        type:
          type: string
          enum: [rules, ml]
    StrategiesResponse:
      type: object
      required: [strategies]
      properties:
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/StrategyMeta"
    StrategyParamSpec:
      type: object
      properties:
        type:
          type: string
          enum: [int, float, string, bool]
        min:
          type: number
          nullable: true
        max:
          type: number
          nullable: true
        default:
          nullable: true
          oneOf:
            - type: number
            - type: string
            - type: boolean
    StrategyDetail:
      allOf:
        - $ref: "#/components/schemas/StrategyMeta"
        - type: object
          required: [parameters_schema]
          properties:
            parameters_schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/StrategyParamSpec"

    # ---- Backtests ----
...